Создал класс аналитики.
